refactor(agent): relocate AgentRuntime + rename createRuntime to src/app.ts
Resolves the long-standing two-runtime naming collision by separating
the agent-layer dispatcher from the application bootstrapper:

- src/legacy/agents/runtime.ts → src/agent/runtime.ts (the AgentRuntime
  class, its 3 test files, and the embedded constants)
- src/legacy/core/runtime.ts → src/app.ts (the createRuntime() process
  bootstrapper, formerly sharing the runtime name)
- src/agent/runtime/types.ts merged into src/agent/types.ts; runtime/
  subfolder removed (file/folder name clash with runtime.ts)
- src/legacy/agents/index.ts deleted (no external consumers)

Internal imports updated across discord/http plugins, slash-commands,
tools, agent-run, session-factory, claude/runner, and cli.

After this PR src/legacy/agents/ contains only builtin/ (the spawn
helper system prompt), to be relocated separately.
