+++ src/agent/types.ts
@@ -1,7 +1,9 @@
 // src/agent/types.ts — Agent-layer types (config, runtime contract)
 
+import type { AgentTool } from "@mariozechner/pi-agent-core";
 import type { SandboxConfig } from "../legacy/sandbox/config.js";
 import type { AgentToolSettings } from "../tools/types.js";
+import type { DefaultSessionStore } from "../legacy/core/session-store.js";
 
 // ---------------------------------------------------------------------------
 // Provider config (single global provider; agents pick model only)
@@ -64,4 +66,59 @@ export interface CompactionConfig {
   preserveRecent?: number;
   /** Absolute token reserve before compaction triggers. Overrides threshold if set. */
   reserveTokens?: number;
+}
+
+// ---------------------------------------------------------------------------
+// Runtime contract — RegisteredAgent / SendMessageRequest / RunInfo
+// ---------------------------------------------------------------------------
+
+export type RunStatus = "created" | "running" | "awaiting" | "completed" | "failed" | "cancelled";
+
+export type AgentSessionKind = "root" | "leaf";
+
+/** "always-new": fresh session per send_message call.
+ *  "parent-reuse": same `(caller, parentSessionId)` reuses one target
+ *  session across calls; falls back to fresh when no parentSessionId. */
+export type AgentSessionPolicy = "always-new" | "parent-reuse";
+
+export interface RegisteredAgent {
+  readonly id: string;
+  config: AgentConfig;
+  readonly sessionStore: DefaultSessionStore;
+  readonly capabilities: {
+    tools: string[];
+    canBeAddressed: boolean;
+  };
+  /** Defaults to "parent-reuse". */
+  readonly sessionPolicy?: AgentSessionPolicy;
+}
+
+export interface SendMessageRequest {
+  to: string;
+  sessionId?: string;
+  content: string;
+  from?: { agentId: string; displayName?: string };
+  parentSessionId?: string;
+  cwd?: string;
+  timeoutSeconds?: number;
+  leafContext?: {
+    /** Parent's filtered tools (parent's tools minus denied for spawn). */
+    tools: AgentTool[];
+    extraSystemPrompt?: string;
+  };
+  /** Fires once after run is registered, before any AgentEvent yields.
+   * Use to wire side-channel UI (Discord thread, audit) by runId. */
+  onRunStart?: (runId: string) => void;
+  /** Fires when `runtime.cancel(runId, { reason })` runs. Lets the caller
+   * shape the LLM-facing result string (e.g. "user cancel — don't retry"). */
+  onCancel?: (reason: string) => void;
+}
+
+export interface RunInfo {
+  runId: string;
+  agentId: string;
+  kind: AgentSessionKind;
+  sessionId: string;
+  startedAt: number;
+  parentSessionId?: string;
 }
