--- src/legacy/agents/index.ts
@@ -1,17 +1,0 @@
-// src/agents/index.ts — Public surface of the agent layer.
-
-export type {
-  RunStatus,
-  AgentSessionKind,
-  RegisteredAgent,
-  SendMessageRequest,
-  RunInfo,
-} from "../../agent/runtime/types.js";
-
-export {
-  AgentRuntime,
-  LEAF_CONCURRENCY_CAP,
-  LEAF_DEFAULT_TIMEOUT_SEC,
-  RESERVED_AGENT_IDS,
-} from "./runtime.js";
-export type { AgentRuntimeOptions } from "./runtime.js";
