+++ src/legacy/core/agent-run.test.ts
@@ -2,8 +2,8 @@
 
 import { describe, it, expect, vi } from "vitest";
 import { consumeRootRun, cancelRunBySessionId } from "./agent-run.js";
-import { AgentRuntime } from "../agents/runtime.js";
-import type { RegisteredAgent, SendMessageRequest } from "../../agent/runtime/types.js";
+import { AgentRuntime } from "../../agent/runtime.js";
+import type { RegisteredAgent, SendMessageRequest } from "../../agent/types.js";
 import type { AgentEvent } from "@mariozechner/pi-agent-core";
 import { createLogger } from "../../logging/logger.js";
