+++ src/legacy/core/agent-run.ts
@@ -2,7 +2,7 @@
 // callers (REST/Discord/heartbeat/cron). Also wires hooks, usage tracker,
 // runtime per-session event emission, and mid-turn steer-from-pending-buffer.
 
-import type { AgentRuntime } from "../agents/runtime.js";
+import type { AgentRuntime } from "../../agent/runtime.js";
 import { userMessage, assistantMessage, getAgentEndMeta } from "../../agent/runners/pi/messages.js";
 import type { Logger } from "../../logging/logger.js";
 import type { HookRegistry } from "../plugins/hooks.js";
