+++ src/legacy/core/tools.ts
@@ -15,9 +15,9 @@ import type { SandboxConfig } from "../sandbox/config.js";
 import { createWebFetchTool, createWebSearchTool } from "../tools/web.js";
 import { createReactTools, type LazyTransportContext } from "../tools/react.js";
 import { createExecTools, ProcessRegistry } from "../tools/exec.js";
-import type { AgentRuntime } from "../agents/runtime.js";
-import { SUBAGENT_AGENT_ID, CLAUDE_AGENT_ID, SendMessageValidationError } from "../agents/runtime.js";
-import type { SendMessageRequest } from "../../agent/runtime/types.js";
+import type { AgentRuntime } from "../../agent/runtime.js";
+import { SUBAGENT_AGENT_ID, CLAUDE_AGENT_ID, SendMessageValidationError } from "../../agent/runtime.js";
+import type { SendMessageRequest } from "../../agent/types.js";
 import { getMessageContext } from "../transport/context.js";
 import { getDiscordSubagentStreamContext } from "../plugins/discord/subagent-stream-context.js";
 import { DiscordSubagentSink } from "../plugins/discord/discord-subagent-sink.js";
