+++ src/legacy/plugins/discord/discord.test.ts
@@ -5,7 +5,7 @@ import { DiscordTransport } from "./discord.js";
 import type { SessionStore } from "../../../sessions/types.js";
 import { ThreadBindingManager } from "./thread-bindings.js";
 import { createMockSessionStore } from "../../core/test-helpers.js";
-import { AgentRuntime } from "../../agents/runtime.js";
+import { AgentRuntime } from "../../../agent/runtime.js";
 
 function makeMockRuntime(agentId: string, cache: unknown, sessionStore: SessionStore): AgentRuntime {
   const rt = new AgentRuntime({ globalProvider: { type: "anthropic", defaultModel: "claude-opus-4.5" } });
