+++ src/legacy/plugins/types.ts
@@ -8,7 +8,7 @@ import type { Transport } from "../../gateway/types.js";
 import type { SessionStoreManager } from "../core/session-store-manager.js";
 import type { IsotopesConfigFile } from "../../config.js";
 import type { LazyTransportContext } from "../tools/react.js";
-import type { AgentRuntime } from "../agents/runtime.js";
+import type { AgentRuntime } from "../../agent/runtime.js";
 
 // ---------------------------------------------------------------------------
 // Plugin manifest (isotopes.plugin.json)
